requirements.txt
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-Cors
Flask-Limiter
Flask-SQLAlchemy
Flask-Migrate
flask-talisman
gunicorn
pandas
numpy
scikit-learn
sentence-transformers
torch
boto3
psycopg2-binary
python-dotenv
pdfminer.six
python-docx
resend
openai
pytest
pytest-cov
